Final changes to the folder structure
