+++ app/src/main/java/ovh/gabrielhuav/sensores_escom_v2/presentation/common/base/GameplayActivity.kt
@@ -23,14 +23,14 @@ import ovh.gabrielhuav.sensores_escom_v2.presentation.common.managers.MovementMa
 import ovh.gabrielhuav.sensores_escom_v2.presentation.common.managers.ServerConnectionManager
 import ovh.gabrielhuav.sensores_escom_v2.presentation.components.BuildingEdificioIA
 import ovh.gabrielhuav.sensores_escom_v2.presentation.components.BuildingNumber2
-import ovh.gabrielhuav.sensores_escom_v2.presentation.components.BuildingNumber4
-import ovh.gabrielhuav.sensores_escom_v2.presentation.components.Cafeteria
-import ovh.gabrielhuav.sensores_escom_v2.presentation.components.EstacionamientoEscom
-import ovh.gabrielhuav.sensores_escom_v2.presentation.components.Zacatenco
-import ovh.gabrielhuav.sensores_escom_v2.presentation.components.ipn.zacatenco.escom.buildingNumber3.SalonPacman
+import ovh.gabrielhuav.sensores_escom_v2.presentation.locations.buildings.building4.BuildingNumber4
+import ovh.gabrielhuav.sensores_escom_v2.presentation.locations.buildings.cafeteria.Cafeteria
+import ovh.gabrielhuav.sensores_escom_v2.presentation.locations.outdoor.EstacionamientoEscom
+import ovh.gabrielhuav.sensores_escom_v2.presentation.locations.outdoor.Zacatenco
+import ovh.gabrielhuav.sensores_escom_v2.presentation.locations.buildings.building3.SalonPacman
 import ovh.gabrielhuav.sensores_escom_v2.presentation.game.mapview.MapMatrixProvider
 import ovh.gabrielhuav.sensores_escom_v2.presentation.game.mapview.MapView
-import ovh.gabrielhuav.sensores_escom_v2.presentation.locations.buildings.PalapasIA
+import ovh.gabrielhuav.sensores_escom_v2.presentation.locations.buildings.buildingIA.PalapasIA
 import kotlin.collections.iterator
 
 class GameplayActivity : AppCompatActivity(),
