+++ app/src/main/java/ovh/gabrielhuav/sensores_escom_v2/presentation/game/zombie/FogOfWarRendere.kt
@@ -1,9 +1,8 @@
-package ovh.gabrielhuav.sensores_escom_v2.presentation.components.ipn.zacatenco.escom.cafeteria
+package ovh.gabrielhuav.sensores_escom_v2.presentation.game.zombie
 
 import android.graphics.Canvas
 import android.graphics.Color
 import android.graphics.Paint
-import android.graphics.Path
 import android.graphics.PorterDuff
 import android.graphics.PorterDuffXfermode
 import android.graphics.RadialGradient
