+++ app/src/main/java/ovh/gabrielhuav/sensores_escom_v2/presentation/locations/buildings/building2/BuildingNumber2Piso1.kt
@@ -1,11 +1,8 @@
 package ovh.gabrielhuav.sensores_escom_v2.presentation.components
 
-import android.Manifest
 import android.bluetooth.BluetoothDevice
 import android.content.Intent
-import android.content.pm.PackageManager
 import android.content.res.Configuration
-import android.net.Uri
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
@@ -15,7 +12,6 @@ import android.widget.FrameLayout
 import android.widget.Toast
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.app.ActivityCompat
 import org.json.JSONObject
 import ovh.gabrielhuav.sensores_escom_v2.R
 import ovh.gabrielhuav.sensores_escom_v2.data.map.Bluetooth.BluetoothGameManager
@@ -25,11 +21,12 @@ import ovh.gabrielhuav.sensores_escom_v2.domain.bluetooth.BluetoothManager
 import ovh.gabrielhuav.sensores_escom_v2.presentation.common.components.UIManager
 import ovh.gabrielhuav.sensores_escom_v2.presentation.common.managers.MovementManager
 import ovh.gabrielhuav.sensores_escom_v2.presentation.common.managers.ServerConnectionManager
-import ovh.gabrielhuav.sensores_escom_v2.presentation.locations.transportation.Cablebus
+import ovh.gabrielhuav.sensores_escom_v2.presentation.components.ipn.zacatenco.escom.buildingNumber2.classrooms.Salon2009
 import ovh.gabrielhuav.sensores_escom_v2.presentation.game.mapview.MapMatrixProvider
 import ovh.gabrielhuav.sensores_escom_v2.presentation.game.mapview.MapView
+import ovh.gabrielhuav.sensores_escom_v2.presentation.common.base.GameplayActivity
 
-class Lindavista : AppCompatActivity(),
+class BuildingNumber2Piso1 : AppCompatActivity(),
     BluetoothManager.BluetoothManagerCallback,
     BluetoothGameManager.ConnectionListener,
     OnlineServerManager.WebSocketListener,
@@ -49,7 +46,7 @@ class Lindavista : AppCompatActivity(),
     data class GameState(
         var isServer: Boolean = false,
         var isConnected: Boolean = false,
-        var playerPosition: Pair<Int, Int> = Pair(10, 12),
+        var playerPosition: Pair<Int, Int> = Pair(1, 1),
         var remotePlayerPositions: Map<String, PlayerInfo> = emptyMap(),
         var remotePlayerName: String? = null
     ) {
@@ -71,13 +68,13 @@ class Lindavista : AppCompatActivity(),
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_lindavista)
+        setContentView(R.layout.activity_building2piso1)
 
         try {
             // Primero inicializamos el mapView
             mapView = MapView(
                 context = this,
-                mapResourceId = R.drawable.lindavista
+                mapResourceId = R.drawable.escom_edificio_2_primera_planta
             )
             findViewById<FrameLayout>(R.id.map_container).addView(mapView)
 
@@ -87,8 +84,8 @@ class Lindavista : AppCompatActivity(),
             // Esperar a que el mapView esté listo
             mapView.post {
                 // Configurar el mapa
-                val normalizedMap = MapMatrixProvider.normalizeMapName(MapMatrixProvider.MAP_LINDAVISTA)
-                mapView.setCurrentMap(normalizedMap, R.drawable.lindavista)
+                val normalizedMap = MapMatrixProvider.normalizeMapName(MapMatrixProvider.MAP_BUILDING2_PISO1)
+                mapView.setCurrentMap(normalizedMap, R.drawable.escom_edificio_2_primera_planta)
 
                 // Después configurar el playerManager
                 mapView.playerManager.apply {
@@ -97,7 +94,7 @@ class Lindavista : AppCompatActivity(),
                     updateLocalPlayerPosition(gameState.playerPosition)
                 }
 
-                Log.d("Lindavista", "Set map to: $normalizedMap")
+                Log.d("BuildingNumber2Piso1", "Set map to: $normalizedMap")
             }
         } catch (e: Exception) {
             Log.e(TAG, "Error en onCreate: ${e.message}")
@@ -117,7 +114,7 @@ class Lindavista : AppCompatActivity(),
         if (savedInstanceState == null) {
             // Inicializar el estado del juego desde el Intent
             gameState.isServer = intent.getBooleanExtra("IS_SERVER", false)
-            gameState.playerPosition = intent.getParcelableExtra("INITIAL_POSITION") ?: Pair(1, 6)
+            gameState.playerPosition = (intent.getParcelableExtra("INITIAL_POSITION") ?: Pair(17, 20)) as Pair<Int, Int>
         } else {
             restoreState(savedInstanceState)
         }
@@ -145,14 +142,14 @@ class Lindavista : AppCompatActivity(),
 
     private fun initializeManagers() {
         bluetoothManager = BluetoothManager.getInstance(this, uiManager.tvBluetoothStatus).apply {
-            setCallback(this@Lindavista)
+            setCallback(this@BuildingNumber2Piso1)
         }
 
         bluetoothBridge = BluetoothWebSocketBridge.getInstance()
 
         // Configurar OnlineServerManager con el listener
         val onlineServerManager = OnlineServerManager.getInstance(this).apply {
-            setListener(this@Lindavista)
+            setListener(this@BuildingNumber2Piso1)
         }
 
         serverConnectionManager = ServerConnectionManager(
@@ -177,15 +174,61 @@ class Lindavista : AppCompatActivity(),
     // Actualiza el método onMapTransitionRequested para manejar la transición al salón 2009
     override fun onMapTransitionRequested(targetMap: String, initialPosition: Pair<Int, Int>) {
         when (targetMap) {
-            MapMatrixProvider.MAP_ZACATENCO -> {
+            MapMatrixProvider.MAP_MAIN -> {
                 // Transición al mapa principal
-                returnToZacatencoActivity()
+                //returnToMainActivity()
+            }
+            MapMatrixProvider.MAP_SALON2009 -> {
+                // Transición al salón 2009
+                //startSalon2009Activity()
+            }
+            MapMatrixProvider.MAP_SALON2010 -> {
+                // Transición al salón 2010
+                //startSalon2010Activity()
+            }
+            MapMatrixProvider.MAP_BUILDING2 -> {
+                // Transición al edificio 2
+                startBuilding2Activity()
             }
             // Añadir más casos según sea necesario para otros mapas
             else -> {
                 Log.d(TAG, "Mapa destino no reconocido: $targetMap")
             }
         }
+    }
+
+    // Método para iniciar la Activity del salón 2010
+    private fun startBuilding2Activity() {
+        val previousPosition = intent.getSerializableExtra("PREVIOUS_POSITION") as? Pair<Int, Int>
+            ?: Pair(17, 20) // Posición por defecto si no hay previa
+
+        val intent = Intent(this, BuildingNumber2::class.java).apply {
+            putExtra("PLAYER_NAME", playerName)
+            putExtra("IS_SERVER", gameState.isServer)
+            putExtra("INITIAL_POSITION", previousPosition) // Usar la posición previa
+            flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TOP
+        }
+
+        // Limpiar datos antes de cambiar de activity
+        mapView.playerManager.cleanup()
+        startActivity(intent)
+        finish()
+    }
+
+    // Método para iniciar la Activity del salón 2009
+    private fun startSalon2009Activity() {
+        val intent = Intent(this, Salon2009::class.java).apply {
+            putExtra("PLAYER_NAME", playerName)
+            putExtra("IS_SERVER", gameState.isServer)
+            putExtra("INITIAL_POSITION", Pair(20, 20)) // Posición inicial en el salón
+            putExtra("PREVIOUS_POSITION", gameState.playerPosition) // Guardar la posición actual para regresar
+            flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TOP
+        }
+
+        // Limpiar datos antes de cambiar de activity
+        mapView.playerManager.cleanup()
+        startActivity(intent)
+        finish()
     }
 
     private fun restoreState(savedInstanceState: Bundle) {
@@ -268,54 +311,34 @@ class Lindavista : AppCompatActivity(),
     }
 
     private var canChangeMap = false  // Variable para controlar si se puede cambiar de mapa
-    private var targetDestination: String? = null  // Variable para almacenar el destino
+    private var targetMapId: String? = null  // Añadir esta variable para almacenar el mapa destino
+    private var interactivePosition: Pair<Int, Int>? = null  // Coordenadas del punto interactivo
 
     private fun checkPositionForMapChange(position: Pair<Int, Int>) {
+        // Verificar si estamos en un punto interactivo que puede ser una transición
+        targetMapId = mapView.getMapTransitionPoint(position.first, position.second)
+        interactivePosition = if (targetMapId != null) position else null
+        canChangeMap = targetMapId != null
 
-        when {
-            position.first == 1 && position.second == 6 -> {
-                canChangeMap = true
-                targetDestination = "zacatenco"
-                runOnUiThread {
-                    Toast.makeText(this, "Presiona A para entrar a Zacatenco", Toast.LENGTH_SHORT)
-                        .show()
+        if (canChangeMap) {
+            runOnUiThread {
+                when (targetMapId) {
+                    MapMatrixProvider.MAP_MAIN -> {
+                        Toast.makeText(this, "Presiona A para volver al mapa principal", Toast.LENGTH_SHORT).show()
+                    }
+                    MapMatrixProvider.MAP_SALON2009 -> {
+                        Toast.makeText(this, "Presiona A para entrar al salón 2009", Toast.LENGTH_SHORT).show()
+                    }
+                    MapMatrixProvider.MAP_SALON2010 -> {
+                        Toast.makeText(this, "Presiona A para entrar al salón 2010", Toast.LENGTH_SHORT).show()
+                    }
+                    MapMatrixProvider.MAP_BUILDING2 -> {
+                        Toast.makeText(this, "Presiona A para entrar al edificio 2 planta baja", Toast.LENGTH_SHORT).show()
+                    }
+                    else -> {
+                        Toast.makeText(this, "Presiona A para interactuar", Toast.LENGTH_SHORT).show()
+                    }
                 }
-            }
-            position.first == 33 && position.second == 34 -> {
-                canChangeMap = true
-                targetDestination = "indios"
-                runOnUiThread {
-                    Toast.makeText(this, "Presiona A para ver Indios Verdes", Toast.LENGTH_SHORT)
-                        .show()
-                }
-            }
-            position.first == 30 && position.second == 9 -> {
-                canChangeMap = true
-                targetDestination = "plaza"
-                runOnUiThread {
-                    Toast.makeText(this, "Presiona A para ver Plaza Vista Norte", Toast.LENGTH_SHORT)
-                        .show()
-                }
-            }
-            position.first == 30 && position.second == 23 -> {
-                canChangeMap = true
-                targetDestination = "talleres"
-                runOnUiThread {
-                    Toast.makeText(this, "Presiona A para ver los Talleres Ticoman", Toast.LENGTH_SHORT)
-                        .show()
-                }
-            }
-            position.first == 26 && position.second == 29 -> {
-                canChangeMap = true
-                targetDestination = "cablebus"
-                runOnUiThread {
-                    Toast.makeText(this, "Presiona A para subir al cablebus", Toast.LENGTH_SHORT)
-                        .show()
-                }
-            }
-            else -> {
-                canChangeMap = false
-                targetDestination = null
             }
         }
     }
@@ -329,7 +352,7 @@ class Lindavista : AppCompatActivity(),
 
             // Añadir el listener para el botón de regreso
             btnConnectDevice.setOnClickListener {
-                returnToZacatencoActivity()
+                returnToMainActivity()
             }
 
             btnNorth.setOnTouchListener { _, event -> handleMovement(event, 0, -1); true }
@@ -339,15 +362,9 @@ class Lindavista : AppCompatActivity(),
 
             // Modificar el botón A para manejar las transiciones de mapa
             buttonA.setOnClickListener {
-                if (canChangeMap) {
-                    when (targetDestination) {
-                        "zacatenco" -> returnToZacatencoActivity()
-                        "indios" -> viewIndiosVerdes()
-                        "plaza" -> viewPlazaVistaNorte()
-                        "talleres" -> viewTalleresTicoman()
-                        "cablebus" -> enterCablebus()
-                        else -> showToast("No hay interacción disponible en esta posición")
-                    }
+                if (canChangeMap && targetMapId != null) {
+                    // En lugar de hacer la lógica aquí directamente, usa el método en MapView
+                    mapView.initiateMapTransition(targetMapId!!)
                 } else {
                     showToast("No hay interacción disponible en esta posición")
                 }
@@ -355,48 +372,19 @@ class Lindavista : AppCompatActivity(),
         }
     }
 
-    private fun viewIndiosVerdes() {
-        val intent = Intent(Intent.ACTION_VIEW, Uri.parse("https://maps.app.goo.gl/r7pcPXFPxNbGcviV8"))
-        startActivity(intent)
-    }
-    private fun viewPlazaVistaNorte() {
-        val intent = Intent(Intent.ACTION_VIEW, Uri.parse("https://maps.app.goo.gl/yuDsYdMcNgQiLfJX9"))
-        startActivity(intent)
-    }
-    private fun viewTalleresTicoman() {
-        val intent = Intent(Intent.ACTION_VIEW, Uri.parse("https://maps.app.goo.gl/g6mWn6vzEZJZsbeb7"))
-        startActivity(intent)
-    }
-
-    private fun returnToZacatencoActivity() {
+    private fun returnToMainActivity() {
         // Obtener la posición previa del intent
         val previousPosition = intent.getSerializableExtra("PREVIOUS_POSITION") as? Pair<Int, Int>
-            ?: Pair(34, 17) // Posición por defecto si no hay previa
+            ?: Pair(15, 10) // Posición por defecto si no hay previa
 
-        val intent = Intent(this, Zacatenco::class.java).apply {
+        val intent = Intent(this, GameplayActivity::class.java).apply {
             putExtra("PLAYER_NAME", playerName)
             putExtra("IS_SERVER", gameState.isServer)
             putExtra("INITIAL_POSITION", previousPosition) // Usar la posición previa
             flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TOP
         }
 
-        // Limpiar datos antes de cambiar de activity
-        mapView.playerManager.cleanup()
-        startActivity(intent)
-        finish()
-    }
 
-    private fun enterCablebus() {
-        // Obtener la posición previa del intent
-        val previousPosition = intent.getSerializableExtra("PREVIOUS_POSITION") as? Pair<Int, Int>
-            ?: Pair(34, 17) // Posición por defecto si no hay previa
-
-        val intent = Intent(this, Cablebus::class.java).apply {
-            putExtra("PLAYER_NAME", playerName)
-            putExtra("IS_SERVER", gameState.isServer)
-            putExtra("INITIAL_POSITION", previousPosition) // Usar la posición previa
-            flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TOP
-        }
 
         // Limpiar datos antes de cambiar de activity
         mapView.playerManager.cleanup()
@@ -413,7 +401,7 @@ class Lindavista : AppCompatActivity(),
                 mapView.updateLocalPlayerPosition(position, forceCenter = true)
 
                 if (gameState.isConnected) {
-                    serverConnectionManager.sendUpdateMessage(playerName, position, "escom_lindavista")
+                    serverConnectionManager.sendUpdateMessage(playerName, position, "escom_building2_primera_planta")
                 }
 
                 checkPositionForMapChange(position)
@@ -439,20 +427,6 @@ class Lindavista : AppCompatActivity(),
 
     // Bluetooth Callbacks
     override fun onBluetoothDeviceConnected(device: BluetoothDevice) {
-        if (ActivityCompat.checkSelfPermission(
-                this,
-                Manifest.permission.BLUETOOTH_CONNECT
-            ) != PackageManager.PERMISSION_GRANTED
-        ) {
-            // TODO: Consider calling
-            //    ActivityCompat#requestPermissions
-            // here to request the missing permissions, and then overriding
-            //   public void onRequestPermissionsResult(int requestCode, String[] permissions,
-            //                                          int[] grantResults)
-            // to handle the case where the user grants the permission. See the documentation
-            // for ActivityCompat#requestPermissions for more details.
-            return
-        }
         gameState.remotePlayerName = device.name
         uiManager.updateBluetoothStatus("Conectado a ${device.name}")
     }
@@ -471,20 +445,6 @@ class Lindavista : AppCompatActivity(),
     }
 
     override fun onDeviceConnected(device: BluetoothDevice) {
-        if (ActivityCompat.checkSelfPermission(
-                this,
-                Manifest.permission.BLUETOOTH_CONNECT
-            ) != PackageManager.PERMISSION_GRANTED
-        ) {
-            // TODO: Consider calling
-            //    ActivityCompat#requestPermissions
-            // here to request the missing permissions, and then overriding
-            //   public void onRequestPermissionsResult(int requestCode, String[] permissions,
-            //                                          int[] grantResults)
-            // to handle the case where the user grants the permission. See the documentation
-            // for ActivityCompat#requestPermissions for more details.
-            return
-        }
         gameState.remotePlayerName = device.name
     }
 
@@ -584,25 +544,9 @@ class Lindavista : AppCompatActivity(),
 
     override fun onPositionReceived(device: BluetoothDevice, x: Int, y: Int) {
         runOnUiThread {
-            val deviceName = if (ActivityCompat.checkSelfPermission(
-                    this,
-                    Manifest.permission.BLUETOOTH_CONNECT
-                ) != PackageManager.PERMISSION_GRANTED
-            ) {
-                // TODO: Consider calling
-                //    ActivityCompat#requestPermissions
-                // here to request the missing permissions, and then overriding
-                //   public void onRequestPermissionsResult(int requestCode, String[] permissions,
-                //                                          int[] grantResults)
-                // to handle the case where the user grants the permission. See the documentation
-                // for ActivityCompat#requestPermissions for more details.
-                return@runOnUiThread
-            } else {
-
-            }
-            device.name ?: "Unknown"
+            val deviceName = device.name ?: "Unknown"
             val currentMap = mapView.playerManager.getCurrentMap()
-            mapView.updateRemotePlayerPosition(deviceName.toString(), Pair(x, y), currentMap)
+            mapView.updateRemotePlayerPosition(deviceName, Pair(x, y), currentMap)
             Log.d("GameplayActivity", "Recibida posición del dispositivo $deviceName: ($x, $y)")
             mapView.invalidate()
         }
