+++ app/src/main/java/ovh/gabrielhuav/sensores_escom_v2/presentation/locations/buildings/building4/BuildingNumber4.kt
@@ -1,11 +1,8 @@
-package ovh.gabrielhuav.sensores_escom_v2.presentation.components
+package ovh.gabrielhuav.sensores_escom_v2.presentation.locations.buildings.building4
 
-import android.Manifest
 import android.bluetooth.BluetoothDevice
 import android.content.Intent
-import android.content.pm.PackageManager
 import android.content.res.Configuration
-import android.net.Uri
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
@@ -15,23 +12,23 @@ import android.widget.FrameLayout
 import android.widget.Toast
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.app.ActivityCompat
 import org.json.JSONObject
 import ovh.gabrielhuav.sensores_escom_v2.R
 import ovh.gabrielhuav.sensores_escom_v2.data.map.Bluetooth.BluetoothGameManager
 import ovh.gabrielhuav.sensores_escom_v2.data.map.Bluetooth.BluetoothWebSocketBridge
 import ovh.gabrielhuav.sensores_escom_v2.data.map.OnlineServer.OnlineServerManager
 import ovh.gabrielhuav.sensores_escom_v2.domain.bluetooth.BluetoothManager
+import ovh.gabrielhuav.sensores_escom_v2.presentation.common.base.GameplayActivity
 import ovh.gabrielhuav.sensores_escom_v2.presentation.common.components.UIManager
 import ovh.gabrielhuav.sensores_escom_v2.presentation.common.managers.MovementManager
 import ovh.gabrielhuav.sensores_escom_v2.presentation.common.managers.ServerConnectionManager
-import ovh.gabrielhuav.sensores_escom_v2.presentation.locations.outdoor.SalidaMetro
+import ovh.gabrielhuav.sensores_escom_v2.presentation.components.ipn.zacatenco.escom.buildingNumber2.classrooms.Salon2009
+import ovh.gabrielhuav.sensores_escom_v2.presentation.components.ipn.zacatenco.escom.buildingNumber2.classrooms.Salon2010
 import ovh.gabrielhuav.sensores_escom_v2.presentation.game.mapview.MapMatrixProvider
 import ovh.gabrielhuav.sensores_escom_v2.presentation.game.mapview.MapView
-import ovh.gabrielhuav.sensores_escom_v2.presentation.common.base.GameplayActivity
+import kotlin.collections.iterator
 
-
-class Zacatenco : AppCompatActivity(),
+class BuildingNumber4 : AppCompatActivity(),
     BluetoothManager.BluetoothManagerCallback,
     BluetoothGameManager.ConnectionListener,
     OnlineServerManager.WebSocketListener,
@@ -51,7 +48,7 @@ class Zacatenco : AppCompatActivity(),
     data class GameState(
         var isServer: Boolean = false,
         var isConnected: Boolean = false,
-        var playerPosition: Pair<Int, Int> = Pair(10, 12),
+        var playerPosition: Pair<Int, Int> = Pair(1, 1),
         var remotePlayerPositions: Map<String, PlayerInfo> = emptyMap(),
         var remotePlayerName: String? = null
     ) {
@@ -73,13 +70,13 @@ class Zacatenco : AppCompatActivity(),
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_zacatenco)
+        setContentView(R.layout.activity_building2)
 
         try {
             // Primero inicializamos el mapView
             mapView = MapView(
                 context = this,
-                mapResourceId = R.drawable.zacatenco
+                mapResourceId = R.drawable.escom_edificio_4_piso_2
             )
             findViewById<FrameLayout>(R.id.map_container).addView(mapView)
 
@@ -89,8 +86,8 @@ class Zacatenco : AppCompatActivity(),
             // Esperar a que el mapView esté listo
             mapView.post {
                 // Configurar el mapa
-                val normalizedMap = MapMatrixProvider.normalizeMapName(MapMatrixProvider.MAP_ZACATENCO)
-                mapView.setCurrentMap(normalizedMap, R.drawable.zacatenco)
+                val normalizedMap = MapMatrixProvider.Companion.normalizeMapName(MapMatrixProvider.Companion.MAP_BUILDING2)
+                mapView.setCurrentMap(normalizedMap, R.drawable.escom_edificio_4_piso_2)
 
                 // Después configurar el playerManager
                 mapView.playerManager.apply {
@@ -99,7 +96,7 @@ class Zacatenco : AppCompatActivity(),
                     updateLocalPlayerPosition(gameState.playerPosition)
                 }
 
-                Log.d("Zacatenco", "Set map to: $normalizedMap")
+                Log.d("BuildingNumber2", "Set map to: $normalizedMap")
             }
         } catch (e: Exception) {
             Log.e(TAG, "Error en onCreate: ${e.message}")
@@ -119,7 +116,7 @@ class Zacatenco : AppCompatActivity(),
         if (savedInstanceState == null) {
             // Inicializar el estado del juego desde el Intent
             gameState.isServer = intent.getBooleanExtra("IS_SERVER", false)
-            gameState.playerPosition = intent.getParcelableExtra("INITIAL_POSITION") ?: Pair(10, 12)
+            gameState.playerPosition = (intent.getParcelableExtra("INITIAL_POSITION") ?: Pair(1, 1)) as Pair<Int, Int>
         } else {
             restoreState(savedInstanceState)
         }
@@ -146,15 +143,15 @@ class Zacatenco : AppCompatActivity(),
     }
 
     private fun initializeManagers() {
-        bluetoothManager = BluetoothManager.getInstance(this, uiManager.tvBluetoothStatus).apply {
-            setCallback(this@Zacatenco)
+        bluetoothManager = BluetoothManager.Companion.getInstance(this, uiManager.tvBluetoothStatus).apply {
+            setCallback(this@BuildingNumber4)
         }
 
-        bluetoothBridge = BluetoothWebSocketBridge.getInstance()
+        bluetoothBridge = BluetoothWebSocketBridge.Companion.getInstance()
 
         // Configurar OnlineServerManager con el listener
-        val onlineServerManager = OnlineServerManager.getInstance(this).apply {
-            setListener(this@Zacatenco)
+        val onlineServerManager = OnlineServerManager.Companion.getInstance(this).apply {
+            setListener(this@BuildingNumber4)
         }
 
         serverConnectionManager = ServerConnectionManager(
@@ -179,15 +176,55 @@ class Zacatenco : AppCompatActivity(),
     // Actualiza el método onMapTransitionRequested para manejar la transición al salón 2009
     override fun onMapTransitionRequested(targetMap: String, initialPosition: Pair<Int, Int>) {
         when (targetMap) {
-            MapMatrixProvider.MAP_MAIN -> {
+            MapMatrixProvider.Companion.MAP_MAIN -> {
                 // Transición al mapa principal
                 returnToMainActivity()
+            }
+            MapMatrixProvider.Companion.MAP_SALON2009 -> {
+                // Transición al salón 2009
+                startSalon2009Activity()
+            }
+            MapMatrixProvider.Companion.MAP_SALON2010 -> {
+                // Transición al salón 2010
+                startSalon2010Activity()
             }
             // Añadir más casos según sea necesario para otros mapas
             else -> {
                 Log.d(TAG, "Mapa destino no reconocido: $targetMap")
             }
         }
+    }
+
+    // Método para iniciar la Activity del salón 2010
+    private fun startSalon2010Activity() {
+        val intent = Intent(this, Salon2010::class.java).apply {
+            putExtra("PLAYER_NAME", playerName)
+            putExtra("IS_SERVER", gameState.isServer)
+            putExtra("INITIAL_POSITION", Pair(20, 20)) // Posición inicial en el salón
+            putExtra("PREVIOUS_POSITION", gameState.playerPosition) // Guardar la posición actual para regresar
+            flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TOP
+        }
+
+        // Limpiar datos antes de cambiar de activity
+        mapView.playerManager.cleanup()
+        startActivity(intent)
+        finish()
+    }
+
+    // Método para iniciar la Activity del salón 2009
+    private fun startSalon2009Activity() {
+        val intent = Intent(this, Salon2009::class.java).apply {
+            putExtra("PLAYER_NAME", playerName)
+            putExtra("IS_SERVER", gameState.isServer)
+            putExtra("INITIAL_POSITION", Pair(20, 20)) // Posición inicial en el salón
+            putExtra("PREVIOUS_POSITION", gameState.playerPosition) // Guardar la posición actual para regresar
+            flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TOP
+        }
+
+        // Limpiar datos antes de cambiar de activity
+        mapView.playerManager.cleanup()
+        startActivity(intent)
+        finish()
     }
 
     private fun restoreState(savedInstanceState: Bundle) {
@@ -270,71 +307,31 @@ class Zacatenco : AppCompatActivity(),
     }
 
     private var canChangeMap = false  // Variable para controlar si se puede cambiar de mapa
-    private var targetDestination: String? = null  // Variable para almacenar el destino
+    private var targetMapId: String? = null  // Añadir esta variable para almacenar el mapa destino
+    private var interactivePosition: Pair<Int, Int>? = null  // Coordenadas del punto interactivo
 
     private fun checkPositionForMapChange(position: Pair<Int, Int>) {
+        // Verificar si estamos en un punto interactivo que puede ser una transición
+        targetMapId = mapView.getMapTransitionPoint(position.first, position.second)
+        interactivePosition = if (targetMapId != null) position else null
+        canChangeMap = targetMapId != null
 
-        when {
-            position.first == 10 && position.second == 12 -> {
-                canChangeMap = true
-                targetDestination = "main"
-                runOnUiThread {
-                    Toast.makeText(this, "Presiona A para entrar a ESCOM", Toast.LENGTH_SHORT)
-                        .show()
+        if (canChangeMap) {
+            runOnUiThread {
+                when (targetMapId) {
+                    MapMatrixProvider.Companion.MAP_MAIN -> {
+                        Toast.makeText(this, "Presiona A para volver al mapa principal", Toast.LENGTH_SHORT).show()
+                    }
+                    MapMatrixProvider.Companion.MAP_SALON2009 -> {
+                        Toast.makeText(this, "Presiona A para entrar al salón 2009", Toast.LENGTH_SHORT).show()
+                    }
+                    MapMatrixProvider.Companion.MAP_SALON2010 -> {
+                        Toast.makeText(this, "Presiona A para entrar al salón 2010", Toast.LENGTH_SHORT).show()
+                    }
+                    else -> {
+                        Toast.makeText(this, "Presiona A para interactuar", Toast.LENGTH_SHORT).show()
+                    }
                 }
-            }
-            position.first == 34 && position.second == 17 -> {
-                canChangeMap = true
-                targetDestination = "lindavista"
-                runOnUiThread {
-                    Toast.makeText(this, "Presiona A para salir a Lindavista", Toast.LENGTH_SHORT)
-                        .show()
-                }
-            }
-            position.first == 25 && position.second == 12 -> {
-                canChangeMap = true
-                targetDestination = "esia"
-                runOnUiThread {
-                    Toast.makeText(this, "Presiona A para ver la ESIA", Toast.LENGTH_SHORT)
-                        .show()
-                }
-            }
-            position.first == 31 && position.second == 17 -> {
-                canChangeMap = true
-                targetDestination = "esfm"
-                runOnUiThread {
-                    Toast.makeText(this, "Presiona A para ver la ESFM", Toast.LENGTH_SHORT)
-                        .show()
-                }
-            }
-            position.first == 8 && position.second == 18 -> {
-                canChangeMap = true
-                targetDestination = "cidetec"
-                runOnUiThread {
-                    Toast.makeText(this, "Presiona A para ver el CIDETEC", Toast.LENGTH_SHORT)
-                        .show()
-                }
-            }
-            position.first == 5 && position.second == 16 -> {
-                canChangeMap = true
-                targetDestination = "cic"
-                runOnUiThread {
-                    Toast.makeText(this, "Presiona A para ver el CIC", Toast.LENGTH_SHORT)
-                        .show()
-                }
-            }
-            position.first == 4 && position.second == 19 -> {
-                canChangeMap = true
-                targetDestination = "ford"
-                runOnUiThread {
-                    Toast.makeText(this, "Presiona A para ver la ford", Toast.LENGTH_SHORT)
-                        .show()
-                }
-            }
-
-            else -> {
-                canChangeMap = false
-                targetDestination = null
             }
         }
     }
@@ -358,48 +355,20 @@ class Zacatenco : AppCompatActivity(),
 
             // Modificar el botón A para manejar las transiciones de mapa
             buttonA.setOnClickListener {
-                if (canChangeMap) {
-                    when (targetDestination) {
-                        "main" -> returnToMainActivity()
-                        "lindavista" -> startLindavistaActivity()
-                        "esia" -> viewESIA()
-                        "esfm" -> viewESFM()
-                        "cidetec" -> viewCIDETEC()
-                        "cic" -> viewCIC()
-                        "ford" -> startFordActivity()
-
-                        else -> showToast("No hay interacción disponible en esta posición")
-                    }
+                if (canChangeMap && targetMapId != null) {
+                    // En lugar de hacer la lógica aquí directamente, usa el método en MapView
+                    mapView.initiateMapTransition(targetMapId!!)
                 } else {
                     showToast("No hay interacción disponible en esta posición")
                 }
             }
         }
     }
-    private fun viewESIA(){
-        val intent = Intent(Intent.ACTION_VIEW, Uri.parse("https://www.esiaz.ipn.mx/"))
-        startActivity(intent)
 
-    }
-    private fun viewESFM(){
-        val intent = Intent(Intent.ACTION_VIEW, Uri.parse("https://www.esfm.ipn.mx/"))
-        startActivity(intent)
-
-    }
-    private fun viewCIDETEC(){
-        val intent = Intent(Intent.ACTION_VIEW, Uri.parse("https://www.cidetec.ipn.mx/"))
-        startActivity(intent)
-
-    }
-    private fun viewCIC(){
-        val intent = Intent(Intent.ACTION_VIEW, Uri.parse("https://www.cic.ipn.mx/"))
-        startActivity(intent)
-
-    }
     private fun returnToMainActivity() {
         // Obtener la posición previa del intent
         val previousPosition = intent.getSerializableExtra("PREVIOUS_POSITION") as? Pair<Int, Int>
-            ?: Pair(11, 4) // Posición por defecto si no hay previa
+            ?: Pair(15, 10) // Posición por defecto si no hay previa
 
         val intent = Intent(this, GameplayActivity::class.java).apply {
             putExtra("PLAYER_NAME", playerName)
@@ -414,24 +383,11 @@ class Zacatenco : AppCompatActivity(),
         finish()
     }
 
-    private fun startLindavistaActivity() {
-        val intent = Intent(this, Lindavista::class.java).apply {
+    private fun startBuildingActivity() {
+        val intent = Intent(this, BuildingNumber4::class.java).apply {
             putExtra("PLAYER_NAME", playerName)
             putExtra("IS_SERVER", gameState.isServer)
-            putExtra("INITIAL_POSITION", Pair(1, 6))
-            putExtra("PREVIOUS_POSITION", gameState.playerPosition) // Guarda la posición actual
-            flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TOP
-        }
-        startActivity(intent)
-        finish()
-    }
-
-    private fun startFordActivity() {
-        val intent = Intent(this, SalidaMetro::class.java).apply {
-            putExtra("PLAYER_NAME", playerName)
-            putExtra("IS_SERVER", gameState.isServer)
-            putExtra("INITIAL_POSITION", Pair(1, 6))
-            putExtra("PREVIOUS_POSITION", gameState.playerPosition) // Guarda la posición actual
+            putExtra("INITIAL_POSITION", Pair(1, 1))
             flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TOP
         }
         startActivity(intent)
@@ -447,7 +403,7 @@ class Zacatenco : AppCompatActivity(),
                 mapView.updateLocalPlayerPosition(position, forceCenter = true)
 
                 if (gameState.isConnected) {
-                    serverConnectionManager.sendUpdateMessage(playerName, position, "escom_zacatenco")
+                    serverConnectionManager.sendUpdateMessage(playerName, position, "escom_building4_floor_2")
                 }
 
                 checkPositionForMapChange(position)
@@ -473,20 +429,6 @@ class Zacatenco : AppCompatActivity(),
 
     // Bluetooth Callbacks
     override fun onBluetoothDeviceConnected(device: BluetoothDevice) {
-        if (ActivityCompat.checkSelfPermission(
-                this,
-                Manifest.permission.BLUETOOTH_CONNECT
-            ) != PackageManager.PERMISSION_GRANTED
-        ) {
-            // TODO: Consider calling
-            //    ActivityCompat#requestPermissions
-            // here to request the missing permissions, and then overriding
-            //   public void onRequestPermissionsResult(int requestCode, String[] permissions,
-            //                                          int[] grantResults)
-            // to handle the case where the user grants the permission. See the documentation
-            // for ActivityCompat#requestPermissions for more details.
-            return
-        }
         gameState.remotePlayerName = device.name
         uiManager.updateBluetoothStatus("Conectado a ${device.name}")
     }
@@ -505,20 +447,6 @@ class Zacatenco : AppCompatActivity(),
     }
 
     override fun onDeviceConnected(device: BluetoothDevice) {
-        if (ActivityCompat.checkSelfPermission(
-                this,
-                Manifest.permission.BLUETOOTH_CONNECT
-            ) != PackageManager.PERMISSION_GRANTED
-        ) {
-            // TODO: Consider calling
-            //    ActivityCompat#requestPermissions
-            // here to request the missing permissions, and then overriding
-            //   public void onRequestPermissionsResult(int requestCode, String[] permissions,
-            //                                          int[] grantResults)
-            // to handle the case where the user grants the permission. See the documentation
-            // for ActivityCompat#requestPermissions for more details.
-            return
-        }
         gameState.remotePlayerName = device.name
     }
 
@@ -618,25 +546,9 @@ class Zacatenco : AppCompatActivity(),
 
     override fun onPositionReceived(device: BluetoothDevice, x: Int, y: Int) {
         runOnUiThread {
-            val deviceName = if (ActivityCompat.checkSelfPermission(
-                    this,
-                    Manifest.permission.BLUETOOTH_CONNECT
-                ) != PackageManager.PERMISSION_GRANTED
-            ) {
-                // TODO: Consider calling
-                //    ActivityCompat#requestPermissions
-                // here to request the missing permissions, and then overriding
-                //   public void onRequestPermissionsResult(int requestCode, String[] permissions,
-                //                                          int[] grantResults)
-                // to handle the case where the user grants the permission. See the documentation
-                // for ActivityCompat#requestPermissions for more details.
-                return@runOnUiThread
-            } else {
-
-            }
-            device.name ?: "Unknown"
+            val deviceName = device.name ?: "Unknown"
             val currentMap = mapView.playerManager.getCurrentMap()
-            mapView.updateRemotePlayerPosition(deviceName.toString(), Pair(x, y), currentMap)
+            mapView.updateRemotePlayerPosition(deviceName, Pair(x, y), currentMap)
             Log.d("GameplayActivity", "Recibida posición del dispositivo $deviceName: ($x, $y)")
             mapView.invalidate()
         }
