+++ app/src/main/java/ovh/gabrielhuav/sensores_escom_v2/presentation/locations/buildings/buildingIA/BuildingEdificioIA_Medio.kt
