+++ app/src/main/java/ovh/gabrielhuav/sensores_escom_v2/presentation/locations/buildings/cafeteria/Cafeteria.kt
@@ -1,4 +1,4 @@
-package ovh.gabrielhuav.sensores_escom_v2.presentation.components
+package ovh.gabrielhuav.sensores_escom_v2.presentation.locations.buildings.cafeteria
 
 import android.bluetooth.BluetoothDevice
 import android.content.Intent
@@ -15,17 +15,18 @@ import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import org.json.JSONObject
 import ovh.gabrielhuav.sensores_escom_v2.R
-import ovh.gabrielhuav.sensores_escom_v2.data.map.Bluetooth.BluetoothWebSocketBridge
 import ovh.gabrielhuav.sensores_escom_v2.data.map.Bluetooth.BluetoothGameManager
+import ovh.gabrielhuav.sensores_escom_v2.data.map.Bluetooth.BluetoothWebSocketBridge
 import ovh.gabrielhuav.sensores_escom_v2.data.map.OnlineServer.OnlineServerManager
 import ovh.gabrielhuav.sensores_escom_v2.domain.bluetooth.BluetoothManager
 import ovh.gabrielhuav.sensores_escom_v2.presentation.common.managers.MovementManager
 import ovh.gabrielhuav.sensores_escom_v2.presentation.common.managers.ServerConnectionManager
-import ovh.gabrielhuav.sensores_escom_v2.presentation.components.ipn.zacatenco.escom.cafeteria.FogOfWarRenderer
-import ovh.gabrielhuav.sensores_escom_v2.presentation.components.ipn.zacatenco.escom.cafeteria.ZombieController
+import ovh.gabrielhuav.sensores_escom_v2.presentation.components.BuildingNumber2
 import ovh.gabrielhuav.sensores_escom_v2.presentation.game.mapview.MapMatrixProvider
 import ovh.gabrielhuav.sensores_escom_v2.presentation.game.mapview.MapView
 import ovh.gabrielhuav.sensores_escom_v2.presentation.game.mapview.PlayerManager
+import ovh.gabrielhuav.sensores_escom_v2.presentation.game.zombie.CafeteriaZombieController
+import ovh.gabrielhuav.sensores_escom_v2.presentation.game.zombie.FogOfWarRenderer
 
 class Cafeteria : AppCompatActivity(),
     BluetoothManager.BluetoothManagerCallback,
@@ -56,18 +57,18 @@ class Cafeteria : AppCompatActivity(),
     private var gameState = BuildingNumber2.GameState()
 
     // Controlador para el minijuego del zombie
-    private lateinit var zombieController: ZombieController
+    private lateinit var cafeteriaZombieController: CafeteriaZombieController
 
     // Estado del minijuego
     private var zombieGameActive = false
     private var playerScore = 0
     private var gameStartTime = 0L
     private val GAME_DURATION_MS = 60000L // 60 segundos de duración del juego
-    private var selectedDifficulty = ZombieController.DIFFICULTY_EASY
+    private var selectedDifficulty = CafeteriaZombieController.Companion.DIFFICULTY_EASY
     private val difficultyNames = mapOf(
-        ZombieController.DIFFICULTY_EASY to "Fácil",
-        ZombieController.DIFFICULTY_MEDIUM to "Medio",
-        ZombieController.DIFFICULTY_HARD to "Difícil"
+        CafeteriaZombieController.Companion.DIFFICULTY_EASY to "Fácil",
+        CafeteriaZombieController.Companion.DIFFICULTY_MEDIUM to "Medio",
+        CafeteriaZombieController.Companion.DIFFICULTY_HARD to "Difícil"
     )
 
     private lateinit var fogOfWarRenderer: FogOfWarRenderer
@@ -92,20 +93,20 @@ class Cafeteria : AppCompatActivity(),
             // Esperar a que el mapView esté listo
             mapView.post {
                 // Configurar el mapa para la cafe de la ESCOM
-                mapView.setCurrentMap(MapMatrixProvider.MAP_CAFETERIA, R.drawable.escom_cafeteria)
+                mapView.setCurrentMap(MapMatrixProvider.Companion.MAP_CAFETERIA, R.drawable.escom_cafeteria)
 
                 // Configurar el playerManager
                 mapView.playerManager.apply {
-                    setCurrentMap(MapMatrixProvider.MAP_CAFETERIA)
+                    setCurrentMap(MapMatrixProvider.Companion.MAP_CAFETERIA)
                     localPlayerId = playerName
                     updateLocalPlayerPosition(gameState.playerPosition)
                 }
 
-                Log.d(TAG, "Set map to: " + MapMatrixProvider.MAP_CAFETERIA)
+                Log.d(TAG, "Set map to: " + MapMatrixProvider.Companion.MAP_CAFETERIA)
 
                 // Importante: Enviar un update inmediato para que otros jugadores sepan dónde estamos
                 if (gameState.isConnected) {
-                    serverConnectionManager.sendUpdateMessage(playerName, gameState.playerPosition, MapMatrixProvider.MAP_CAFETERIA)
+                    serverConnectionManager.sendUpdateMessage(playerName, gameState.playerPosition, MapMatrixProvider.Companion.MAP_CAFETERIA)
                 }
 
                 // Mostrar diálogo de bienvenida al minijuego
@@ -138,9 +139,9 @@ class Cafeteria : AppCompatActivity(),
             .setPositiveButton("Cambiar") { dialog, _ ->
                 // Actualizar dificultad
                 selectedDifficulty = when(selectedOption) {
-                    1 -> ZombieController.DIFFICULTY_MEDIUM
-                    2 -> ZombieController.DIFFICULTY_HARD
-                    else -> ZombieController.DIFFICULTY_EASY
+                    1 -> CafeteriaZombieController.Companion.DIFFICULTY_MEDIUM
+                    2 -> CafeteriaZombieController.Companion.DIFFICULTY_HARD
+                    else -> CafeteriaZombieController.Companion.DIFFICULTY_EASY
                 }
 
                 // Reiniciar el juego con la nueva dificultad
@@ -183,11 +184,15 @@ class Cafeteria : AppCompatActivity(),
         updatePlayerPosition(gameState.playerPosition)
 
         // Inicializar el controlador del zombie
-        zombieController = ZombieController(
+        cafeteriaZombieController = CafeteriaZombieController(
             onZombiePositionChanged = { zombieId, position ->
                 // Cuando el zombie cambia de posición, lo dibujamos en el mapa
                 runOnUiThread {
-                    mapView.updateSpecialEntity(zombieId, position, MapMatrixProvider.MAP_CAFETERIA)
+                    mapView.updateSpecialEntity(
+                        zombieId,
+                        position,
+                        MapMatrixProvider.Companion.MAP_CAFETERIA
+                    )
                     mapView.invalidate()
                 }
             },
@@ -230,9 +235,9 @@ class Cafeteria : AppCompatActivity(),
             .setSingleChoiceItems(options, selectedOption) { _, which ->
                 selectedOption = which
                 selectedDifficulty = when(which) {
-                    1 -> ZombieController.DIFFICULTY_MEDIUM
-                    2 -> ZombieController.DIFFICULTY_HARD
-                    else -> ZombieController.DIFFICULTY_EASY
+                    1 -> CafeteriaZombieController.Companion.DIFFICULTY_MEDIUM
+                    2 -> CafeteriaZombieController.Companion.DIFFICULTY_HARD
+                    else -> CafeteriaZombieController.Companion.DIFFICULTY_EASY
                 }
             }
             .setPositiveButton("¡Entendido!") { dialog, _ ->
@@ -259,7 +264,7 @@ class Cafeteria : AppCompatActivity(),
         }
 
         // Iniciar el controlador del zombie con la dificultad seleccionada
-        zombieController.startGame(selectedDifficulty)
+        cafeteriaZombieController.startGame(selectedDifficulty)
 
         // Iniciar timer de actualización
         startGameUpdateTimer()
@@ -284,7 +289,7 @@ class Cafeteria : AppCompatActivity(),
         mapView.setFogOfWarEnabled(false)
 
         if (zombieGameActive) {
-            zombieController.stopGame()
+            cafeteriaZombieController.stopGame()
             zombieGameActive = false
 
             // Detener timer
@@ -323,7 +328,7 @@ class Cafeteria : AppCompatActivity(),
 
     private fun completeZombieGame(survived: Boolean) {
         zombieGameActive = false
-        zombieController.stopGame()
+        cafeteriaZombieController.stopGame()
 
         // Detener timer
         stopGameUpdateTimer()
@@ -360,13 +365,13 @@ class Cafeteria : AppCompatActivity(),
         sendZombieGameUpdate("complete", survived, secondsSurvived.toInt(), playerScore)
     }
 
-    private fun sendZombieGameUpdate(action: String, survived: Boolean = false, time: Int = 0, score: Int = 0, difficulty: Int = ZombieController.DIFFICULTY_EASY) {
+    private fun sendZombieGameUpdate(action: String, survived: Boolean = false, time: Int = 0, score: Int = 0, difficulty: Int = CafeteriaZombieController.Companion.DIFFICULTY_EASY) {
         try {
             val message = JSONObject().apply {
                 put("type", "zombie_game_update")
                 put("action", action)
                 put("player", playerName)
-                put("map", MapMatrixProvider.MAP_CAFETERIA)
+                put("map", MapMatrixProvider.Companion.MAP_CAFETERIA)
 
                 if (action == "start") {
                     put("difficulty", difficulty)
@@ -388,15 +393,15 @@ class Cafeteria : AppCompatActivity(),
     // 7. Añadir método para actualizar las posiciones de los jugadores remotos en el zombieController
     private fun updateRemotePlayersForZombies() {
         // Esta función debe ser llamada cuando se detecten cambios en las posiciones de jugadores remotos
-        if (zombieGameActive && this::zombieController.isInitialized) {
+        if (zombieGameActive && this::cafeteriaZombieController.isInitialized) {
             // Actualizar la posición del jugador local
-            zombieController.updatePlayerPosition(playerName, gameState.playerPosition)
+            cafeteriaZombieController.updatePlayerPosition(playerName, gameState.playerPosition)
 
             // Actualizar las posiciones de jugadores remotos
             gameState.remotePlayerPositions.forEach { (playerId, playerInfo) ->
                 // Solo considerar jugadores en el mismo mapa (la cafetería)
-                if (MapMatrixProvider.normalizeMapName(playerInfo.map) == MapMatrixProvider.MAP_CAFETERIA) {
-                    zombieController.updatePlayerPosition(playerId, playerInfo.position)
+                if (MapMatrixProvider.Companion.normalizeMapName(playerInfo.map) == MapMatrixProvider.Companion.MAP_CAFETERIA) {
+                    cafeteriaZombieController.updatePlayerPosition(playerId, playerInfo.position)
                 }
             }
         }
@@ -422,7 +427,7 @@ class Cafeteria : AppCompatActivity(),
                 Toast.makeText(this, "¡Recogiste comida! +10 puntos", Toast.LENGTH_SHORT).show()
 
                 // Ralentizar a los zombies
-                zombieController.slowDownZombies(3000) // 3 segundos más lentos
+                cafeteriaZombieController.slowDownZombies(3000) // 3 segundos más lentos
 
                 // Informar al servidor
                 val message = JSONObject().apply {
@@ -452,9 +457,9 @@ class Cafeteria : AppCompatActivity(),
             .setSingleChoiceItems(options, selectedOption) { _, which ->
                 selectedOption = which
                 selectedDifficulty = when(which) {
-                    1 -> ZombieController.DIFFICULTY_MEDIUM
-                    2 -> ZombieController.DIFFICULTY_HARD
-                    else -> ZombieController.DIFFICULTY_EASY
+                    1 -> CafeteriaZombieController.Companion.DIFFICULTY_MEDIUM
+                    2 -> CafeteriaZombieController.Companion.DIFFICULTY_HARD
+                    else -> CafeteriaZombieController.Companion.DIFFICULTY_EASY
                 }
             }
             .setPositiveButton("Aceptar") { dialog, _ ->
@@ -483,7 +488,7 @@ class Cafeteria : AppCompatActivity(),
                     serverConnectionManager.sendUpdateMessage(
                         playerName,
                         gameState.playerPosition,
-                        MapMatrixProvider.MAP_CAFETERIA
+                        MapMatrixProvider.Companion.MAP_CAFETERIA
                     )
 
                     // Solicitar actualizaciones de posición
@@ -539,20 +544,20 @@ class Cafeteria : AppCompatActivity(),
     private fun getPlayerVisionRadius(): Int {
         // Usar el mismo rango de detección que los zombies en dificultad difícil
         return when (selectedDifficulty) {
-            ZombieController.DIFFICULTY_HARD -> 18
-            ZombieController.DIFFICULTY_MEDIUM -> 12
+            CafeteriaZombieController.Companion.DIFFICULTY_HARD -> 18
+            CafeteriaZombieController.Companion.DIFFICULTY_MEDIUM -> 12
             else -> 8
         }
     }
 
     private fun initializeManagers() {
-        bluetoothManager = BluetoothManager.getInstance(this, tvBluetoothStatus).apply {
+        bluetoothManager = BluetoothManager.Companion.getInstance(this, tvBluetoothStatus).apply {
             setCallback(this@Cafeteria)
         }
 
-        bluetoothBridge = BluetoothWebSocketBridge.getInstance()
+        bluetoothBridge = BluetoothWebSocketBridge.Companion.getInstance()
 
-        val onlineServerManager = OnlineServerManager.getInstance(this).apply {
+        val onlineServerManager = OnlineServerManager.Companion.getInstance(this).apply {
             setListener(this@Cafeteria)
         }
 
@@ -672,12 +677,12 @@ class Cafeteria : AppCompatActivity(),
             // Enviar actualización a otros jugadores con el mapa específico
             if (gameState.isConnected) {
                 // Enviar la posición con el nombre del mapa correcto
-                serverConnectionManager.sendUpdateMessage(playerName, position, MapMatrixProvider.MAP_CAFETERIA)
+                serverConnectionManager.sendUpdateMessage(playerName, position, MapMatrixProvider.Companion.MAP_CAFETERIA)
             }
 
             // Si el minijuego está activo, actualizar la posición del jugador para el zombie
-            if (zombieGameActive && this::zombieController.isInitialized) {
-                zombieController.updatePlayerPosition(playerName, position)
+            if (zombieGameActive && this::cafeteriaZombieController.isInitialized) {
+                cafeteriaZombieController.updatePlayerPosition(playerName, position)
             }
         }
         if (fogOfWarEnabled && zombieGameActive) {
@@ -709,7 +714,7 @@ class Cafeteria : AppCompatActivity(),
 
         // Restaurar el minijuego si estaba activo
         if (zombieGameActive) {
-            zombieController.startGame()
+            cafeteriaZombieController.startGame()
         }
     }
 
@@ -764,7 +769,7 @@ class Cafeteria : AppCompatActivity(),
     // Implementación MapTransitionListener
     override fun onMapTransitionRequested(targetMap: String, initialPosition: Pair<Int, Int>) {
         when (targetMap) {
-            MapMatrixProvider.MAP_BUILDING2 -> {
+            MapMatrixProvider.Companion.MAP_BUILDING2 -> {
                 returnToBuilding2()
             }
             else -> {
@@ -799,7 +804,7 @@ class Cafeteria : AppCompatActivity(),
     override fun onPositionReceived(device: BluetoothDevice, x: Int, y: Int) {
         runOnUiThread {
             val deviceName = device.name ?: "Unknown"
-            mapView.updateRemotePlayerPosition(deviceName, Pair(x, y), MapMatrixProvider.MAP_CAFETERIA)
+            mapView.updateRemotePlayerPosition(deviceName, Pair(x, y), MapMatrixProvider.Companion.MAP_CAFETERIA)
             mapView.invalidate()
         }
     }
@@ -809,7 +814,7 @@ class Cafeteria : AppCompatActivity(),
         val debugPosition = Pair(20, 20)
 
         // Actualizar la entidad en el PlayerManager directamente
-        mapView.playerManager.updateSpecialEntity("zombie", debugPosition, MapMatrixProvider.MAP_CAFETERIA)
+        mapView.playerManager.updateSpecialEntity("zombie", debugPosition, MapMatrixProvider.Companion.MAP_CAFETERIA)
 
         // Forzar el redibujado
         mapView.invalidate()
@@ -825,18 +830,18 @@ class Cafeteria : AppCompatActivity(),
         mapView.playerManager.logSpecialEntities()
 
         // Si hay un ZombieController inicializado, también actualizar ahí
-        if (this::zombieController.isInitialized) {
+        if (this::cafeteriaZombieController.isInitialized) {
             // Cambiar esta línea:
             // zombieController.setZombiePosition(debugPosition)
             // Por esta:
-            zombieController.setZombiePosition("zombie", debugPosition)
+            cafeteriaZombieController.setZombiePosition("zombie", debugPosition)
         }
     }
 
     private fun forceZombieGameWithFixedZombie() {
         // 1. Detener cualquier juego activo
         if (zombieGameActive) {
-            zombieController.stopGame()
+            cafeteriaZombieController.stopGame()
         }
 
         // 2. Establecer el estado del juego
@@ -857,10 +862,10 @@ class Cafeteria : AppCompatActivity(),
         // Cambiar esta línea:
         // zombieController.setZombiePosition(zombiePosition)
         // Por esta:
-        zombieController.setZombiePosition("zombie", zombiePosition)
+        cafeteriaZombieController.setZombiePosition("zombie", zombiePosition)
 
         // 6. Actualizar la entidad especial en el mapa (esto debe llamar a updateSpecialEntity del PlayerManager)
-        mapView.playerManager.updateSpecialEntity("zombie", zombiePosition, MapMatrixProvider.MAP_CAFETERIA)
+        mapView.playerManager.updateSpecialEntity("zombie", zombiePosition, MapMatrixProvider.Companion.MAP_CAFETERIA)
 
         // 7. Forzar un redibujado inmediato
         mapView.invalidate()
@@ -897,15 +902,15 @@ class Cafeteria : AppCompatActivity(),
 
                                 // Obtener el mapa y normalizarlo
                                 val mapStr = playerData.optString("map", playerData.optString("currentMap", "main"))
-                                val normalizedMap = MapMatrixProvider.normalizeMapName(mapStr)
+                                val normalizedMap = MapMatrixProvider.Companion.normalizeMapName(mapStr)
 
                                 // Guardar en el estado del juego
                                 gameState.remotePlayerPositions = gameState.remotePlayerPositions +
                                         (playerId to BuildingNumber2.GameState.PlayerInfo(position, normalizedMap))
 
                                 // Obtener el mapa actual normalizado para comparar
-                                val currentMap = MapMatrixProvider.normalizeMapName(
-                                    MapMatrixProvider.MAP_CAFETERIA)
+                                val currentMap = MapMatrixProvider.Companion.normalizeMapName(
+                                    MapMatrixProvider.Companion.MAP_CAFETERIA)
 
                                 // Solo mostrar jugadores en el mismo mapa
                                 if (normalizedMap == currentMap) {
@@ -925,14 +930,15 @@ class Cafeteria : AppCompatActivity(),
 
                             // Obtener y normalizar el mapa
                             val mapStr = jsonObject.optString("map", jsonObject.optString("currentmap", "main"))
-                            val normalizedMap = MapMatrixProvider.normalizeMapName(mapStr)
+                            val normalizedMap = MapMatrixProvider.Companion.normalizeMapName(mapStr)
 
                             // Guardar en el estado
                             gameState.remotePlayerPositions = gameState.remotePlayerPositions +
                                     (playerId to BuildingNumber2.GameState.PlayerInfo(position, normalizedMap))
 
                             // Obtener el mapa actual normalizado para comparar
-                            val currentMap = MapMatrixProvider.normalizeMapName(MapMatrixProvider.MAP_CAFETERIA)
+                            val currentMap = MapMatrixProvider.Companion.normalizeMapName(
+                                MapMatrixProvider.Companion.MAP_CAFETERIA)
 
                             // IMPORTANTE: Loggear para depuración
                             Log.d(TAG, "Jugador remoto $playerId en mapa '$normalizedMap', mapa actual es '$currentMap'")
@@ -954,12 +960,12 @@ class Cafeteria : AppCompatActivity(),
                         Log.d(TAG, "¡Recibida posición de zombie $zombieId en ($x, $y)!")
 
                         // Actualizar la posición del zombie en el controlador
-                        if (this::zombieController.isInitialized) {
-                            zombieController.setZombiePosition(zombieId, zombiePosition)
+                        if (this::cafeteriaZombieController.isInitialized) {
+                            cafeteriaZombieController.setZombiePosition(zombieId, zombiePosition)
                         }
 
                         // IMPORTANTE: Actualizar la entidad especial en el mapa
-                        mapView.updateSpecialEntity(zombieId, zombiePosition, MapMatrixProvider.MAP_CAFETERIA)
+                        mapView.updateSpecialEntity(zombieId, zombiePosition, MapMatrixProvider.Companion.MAP_CAFETERIA)
                         mapView.invalidate() // Forzar redibujado
                     }
 
@@ -1018,7 +1024,7 @@ class Cafeteria : AppCompatActivity(),
                         serverConnectionManager.sendUpdateMessage(
                             playerName,
                             gameState.playerPosition,
-                            MapMatrixProvider.MAP_CAFETERIA
+                            MapMatrixProvider.Companion.MAP_CAFETERIA
                         )
                     }
                     "disconnect" -> {
@@ -1040,11 +1046,11 @@ class Cafeteria : AppCompatActivity(),
     private fun normalizeMapId(mapId: String): String {
         // Casos específicos conocidos
         return when {
-            mapId.contains("cafeteria") -> MapMatrixProvider.MAP_CAFETERIA
-            mapId.contains("salon2009") -> MapMatrixProvider.MAP_SALON2009
-            mapId.contains("salon2010") -> MapMatrixProvider.MAP_SALON2010
-            mapId.contains("building2") -> MapMatrixProvider.MAP_BUILDING2
-            mapId.contains("main") -> MapMatrixProvider.MAP_MAIN
+            mapId.contains("cafeteria") -> MapMatrixProvider.Companion.MAP_CAFETERIA
+            mapId.contains("salon2009") -> MapMatrixProvider.Companion.MAP_SALON2009
+            mapId.contains("salon2010") -> MapMatrixProvider.Companion.MAP_SALON2010
+            mapId.contains("building2") -> MapMatrixProvider.Companion.MAP_BUILDING2
+            mapId.contains("main") -> MapMatrixProvider.Companion.MAP_MAIN
             else -> mapId
         }
     }
@@ -1090,7 +1096,7 @@ class Cafeteria : AppCompatActivity(),
             serverConnectionManager.sendUpdateMessage(
                 playerName,
                 gameState.playerPosition,
-                MapMatrixProvider.MAP_CAFETERIA
+                MapMatrixProvider.Companion.MAP_CAFETERIA
             )
         }
     }
