+++ app/src/main/java/ovh/gabrielhuav/sensores_escom_v2/presentation/locations/outdoor/EstacionamientoEscom.kt
@@ -1,4 +1,4 @@
-package ovh.gabrielhuav.sensores_escom_v2.presentation.components
+package ovh.gabrielhuav.sensores_escom_v2.presentation.locations.outdoor
 
 import android.bluetooth.BluetoothDevice
 import android.content.Intent
@@ -16,12 +16,12 @@ import ovh.gabrielhuav.sensores_escom_v2.data.map.Bluetooth.BluetoothGameManager
 import ovh.gabrielhuav.sensores_escom_v2.data.map.Bluetooth.BluetoothWebSocketBridge
 import ovh.gabrielhuav.sensores_escom_v2.data.map.OnlineServer.OnlineServerManager
 import ovh.gabrielhuav.sensores_escom_v2.domain.bluetooth.BluetoothManager
+import ovh.gabrielhuav.sensores_escom_v2.presentation.common.base.GameplayActivity
 import ovh.gabrielhuav.sensores_escom_v2.presentation.common.managers.MovementManager
 import ovh.gabrielhuav.sensores_escom_v2.presentation.common.managers.ServerConnectionManager
+import ovh.gabrielhuav.sensores_escom_v2.presentation.components.BuildingNumber2
 import ovh.gabrielhuav.sensores_escom_v2.presentation.game.mapview.MapMatrixProvider
 import ovh.gabrielhuav.sensores_escom_v2.presentation.game.mapview.MapView
-import ovh.gabrielhuav.sensores_escom_v2.presentation.common.base.GameplayActivity
-import ovh.gabrielhuav.sensores_escom_v2.presentation.locations.outdoor.TramoAtrasPlaza
 
 /**
  * Activity para el mapa del Estacionamiento de ESCOM
@@ -72,20 +72,20 @@ class EstacionamientoEscom : AppCompatActivity(),
             // Esperar a que el mapView esté listo
             mapView.post {
                 // Configurar el mapa para el estacionamiento
-                mapView.setCurrentMap(MapMatrixProvider.MAP_ESTACIONAMIENTO, R.drawable.estacionamiento_escom)
+                mapView.setCurrentMap(MapMatrixProvider.Companion.MAP_ESTACIONAMIENTO, R.drawable.estacionamiento_escom)
 
                 // Configurar el playerManager
                 mapView.playerManager.apply {
-                    setCurrentMap(MapMatrixProvider.MAP_ESTACIONAMIENTO)
+                    setCurrentMap(MapMatrixProvider.Companion.MAP_ESTACIONAMIENTO)
                     localPlayerId = playerName
                     updateLocalPlayerPosition(gameState.playerPosition)
                 }
 
-                Log.d(TAG, "Set map to: " + MapMatrixProvider.MAP_ESTACIONAMIENTO)
+                Log.d(TAG, "Set map to: " + MapMatrixProvider.Companion.MAP_ESTACIONAMIENTO)
 
                 // Importante: Enviar un update inmediato para que otros jugadores sepan dónde estamos
                 if (gameState.isConnected) {
-                    serverConnectionManager.sendUpdateMessage(playerName, gameState.playerPosition, MapMatrixProvider.MAP_ESTACIONAMIENTO)
+                    serverConnectionManager.sendUpdateMessage(playerName, gameState.playerPosition, MapMatrixProvider.Companion.MAP_ESTACIONAMIENTO)
                 }
             }
         } catch (e: Exception) {
@@ -142,7 +142,7 @@ class EstacionamientoEscom : AppCompatActivity(),
                     serverConnectionManager.sendUpdateMessage(
                         playerName,
                         gameState.playerPosition,
-                        MapMatrixProvider.MAP_ESTACIONAMIENTO
+                        MapMatrixProvider.Companion.MAP_ESTACIONAMIENTO
                     )
 
                     // Solicitar actualizaciones de posición
@@ -173,13 +173,13 @@ class EstacionamientoEscom : AppCompatActivity(),
     }
 
     private fun initializeManagers() {
-        bluetoothManager = BluetoothManager.getInstance(this, tvBluetoothStatus).apply {
+        bluetoothManager = BluetoothManager.Companion.getInstance(this, tvBluetoothStatus).apply {
             setCallback(this@EstacionamientoEscom)
         }
 
-        bluetoothBridge = BluetoothWebSocketBridge.getInstance()
+        bluetoothBridge = BluetoothWebSocketBridge.Companion.getInstance()
 
-        val onlineServerManager = OnlineServerManager.getInstance(this).apply {
+        val onlineServerManager = OnlineServerManager.Companion.getInstance(this).apply {
             setListener(this@EstacionamientoEscom)
         }
 
@@ -288,7 +288,7 @@ class EstacionamientoEscom : AppCompatActivity(),
             // Enviar actualización a otros jugadores con el mapa específico
             if (gameState.isConnected) {
                 // Enviar la posición con el nombre del mapa correcto
-                serverConnectionManager.sendUpdateMessage(playerName, position, MapMatrixProvider.MAP_ESTACIONAMIENTO)
+                serverConnectionManager.sendUpdateMessage(playerName, position, MapMatrixProvider.Companion.MAP_ESTACIONAMIENTO)
             }
 
             // Comprobar si estamos en un punto de transición
@@ -302,10 +302,10 @@ class EstacionamientoEscom : AppCompatActivity(),
         if (transitionMap != null) {
             // Mostrar información sobre la transición disponible
             when (transitionMap) {
-                MapMatrixProvider.MAP_MAIN -> {
+                MapMatrixProvider.Companion.MAP_MAIN -> {
                     Toast.makeText(this, "Presiona A para volver al Mapa Principal", Toast.LENGTH_SHORT).show()
                 }
-                MapMatrixProvider.MAP_TRAS_PLAZA -> {
+                MapMatrixProvider.Companion.MAP_TRAS_PLAZA -> {
                     Toast.makeText(this, "Presiona A para ir al Tramo Atrás Plaza", Toast.LENGTH_SHORT).show()
                 }
             }
@@ -333,10 +333,10 @@ class EstacionamientoEscom : AppCompatActivity(),
     // Implementación MapTransitionListener
     override fun onMapTransitionRequested(targetMap: String, initialPosition: Pair<Int, Int>) {
         when (targetMap) {
-            MapMatrixProvider.MAP_MAIN -> {
+            MapMatrixProvider.Companion.MAP_MAIN -> {
                 returnToMainMap()
             }
-            MapMatrixProvider.MAP_TRAS_PLAZA -> {
+            MapMatrixProvider.Companion.MAP_TRAS_PLAZA -> {
                 startNextMap()
             }
             else -> {
@@ -371,7 +371,7 @@ class EstacionamientoEscom : AppCompatActivity(),
     override fun onPositionReceived(device: BluetoothDevice, x: Int, y: Int) {
         runOnUiThread {
             val deviceName = device.name ?: "Unknown"
-            mapView.updateRemotePlayerPosition(deviceName, Pair(x, y), MapMatrixProvider.MAP_ESTACIONAMIENTO)
+            mapView.updateRemotePlayerPosition(deviceName, Pair(x, y), MapMatrixProvider.Companion.MAP_ESTACIONAMIENTO)
             mapView.invalidate()
         }
     }
@@ -396,15 +396,15 @@ class EstacionamientoEscom : AppCompatActivity(),
 
                                 // Obtener y normalizar el mapa
                                 val mapStr = playerData.optString("map", playerData.optString("currentMap", "main"))
-                                val normalizedMap = MapMatrixProvider.normalizeMapName(mapStr)
+                                val normalizedMap = MapMatrixProvider.Companion.normalizeMapName(mapStr)
 
                                 // Actualizar el estado
                                 gameState.remotePlayerPositions = gameState.remotePlayerPositions +
                                         (playerId to BuildingNumber2.GameState.PlayerInfo(position, normalizedMap))
 
                                 // Obtener el mapa actual normalizado para comparar
-                                val currentMap = MapMatrixProvider.normalizeMapName(
-                                    MapMatrixProvider.MAP_ESTACIONAMIENTO)
+                                val currentMap = MapMatrixProvider.Companion.normalizeMapName(
+                                    MapMatrixProvider.Companion.MAP_ESTACIONAMIENTO)
 
                                 // Solo mostrar jugadores en el mismo mapa
                                 if (normalizedMap == currentMap) {
@@ -424,14 +424,15 @@ class EstacionamientoEscom : AppCompatActivity(),
 
                             // Obtener y normalizar el mapa
                             val mapStr = jsonObject.optString("map", jsonObject.optString("currentmap", "main"))
-                            val normalizedMap = MapMatrixProvider.normalizeMapName(mapStr)
+                            val normalizedMap = MapMatrixProvider.Companion.normalizeMapName(mapStr)
 
                             // Actualizar el estado
                             gameState.remotePlayerPositions = gameState.remotePlayerPositions +
                                     (playerId to BuildingNumber2.GameState.PlayerInfo(position, normalizedMap))
 
                             // Obtener el mapa actual normalizado para comparar
-                            val currentMap = MapMatrixProvider.normalizeMapName(MapMatrixProvider.MAP_ESTACIONAMIENTO)
+                            val currentMap = MapMatrixProvider.Companion.normalizeMapName(
+                                MapMatrixProvider.Companion.MAP_ESTACIONAMIENTO)
 
                             // Solo mostrar jugadores en el mismo mapa
                             if (normalizedMap == currentMap) {
@@ -448,7 +449,7 @@ class EstacionamientoEscom : AppCompatActivity(),
                         serverConnectionManager.sendUpdateMessage(
                             playerName,
                             gameState.playerPosition,
-                            MapMatrixProvider.MAP_ESTACIONAMIENTO
+                            MapMatrixProvider.Companion.MAP_ESTACIONAMIENTO
                         )
                     }
                     "disconnect" -> {
@@ -503,7 +504,7 @@ class EstacionamientoEscom : AppCompatActivity(),
             serverConnectionManager.sendUpdateMessage(
                 playerName,
                 gameState.playerPosition,
-                MapMatrixProvider.MAP_ESTACIONAMIENTO
+                MapMatrixProvider.Companion.MAP_ESTACIONAMIENTO
             )
         }
     }
