+++ app/src/main/res/layout/activity_building4.xml
@@ -4,7 +4,7 @@
     android:id="@+id/main_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.components.BuildingNumber4">
+    tools:context=".presentation.locations.buildings.building4.BuildingNumber4">
 
     <Button
         android:id="@+id/button_back_to_home"
