+++ app/src/main/res/layout/activity_estacionamientoescom.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.components.EstacionamientoEscom">
+    tools:context=".presentation.locations.outdoor.EstacionamientoEscom">
 
     <Button
         android:id="@+id/button_back_to_home"
