+++ app/src/main/res/layout/activity_salon_pacman.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.components.ipn.zacatenco.escom.buildingNumber3.SalonPacman">
+    tools:context=".presentation.locations.buildings.building3.SalonPacman">
 
     <Button
         android:id="@+id/button_back_to_home"
